Refactor to support Open Core by making internal packages public
Renamed all 'internal/' directories across the repository (in nexus-broker,
nexus-gateway, and nexus-bridge) to 'pkg/' so they can be imported by
external modules. Updated all corresponding import statements throughout
the codebase to reflect the new 'pkg/' paths instead of 'internal/'.
Ensured code compiles and all tests pass after changes.
