+++ docs/services/bridge.md
@@ -30,7 +30,7 @@ The Bridge includes built-in support for:
 ## Usage Example (HTTP)
 
 ```go
-import "github.com/Prescott-Data/nexus-framework/nexus-bridge/internal/auth"
+import "github.com/Prescott-Data/nexus-framework/nexus-bridge/pkg/auth"
 
 // ... inside your agent ...
 token, _ := sdkClient.GetToken(ctx, connectionID)
