+++ nexus-bridge/integration_test.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/Prescott-Data/nexus-framework/nexus-sdk"
 	"github.com/Prescott-Data/nexus-framework/nexus-bridge"
-	"github.com/Prescott-Data/nexus-framework/nexus-bridge/internal/auth"
+	"github.com/Prescott-Data/nexus-framework/nexus-bridge/pkg/auth"
 	"github.com/gorilla/websocket"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
