+++ nexus-broker/cmd/nexus-broker/main.go
@@ -10,10 +10,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/caching"
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/handlers"
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/provider"
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/server"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/caching"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/handlers"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/provider"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/server"
 	"github.com/go-chi/chi/v5"
 	"github.com/go-redis/redis/v8"
 	"github.com/jmoiron/sqlx"
