+++ nexus-broker/pkg/handlers/callback_test.go
@@ -10,8 +10,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/auth"
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/vault"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/auth"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/vault"
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
 	"github.com/stretchr/testify/assert"
