+++ nexus-broker/pkg/handlers/providers_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 
-	"github.com/Prescott-Data/nexus-framework/nexus-broker/internal/provider"
+	"github.com/Prescott-Data/nexus-framework/nexus-broker/pkg/provider"
 )
 
 // MockStore is a mock implementation of the provider.ProfileStorer interface.
