+++ nexus-gateway/pkg/usecase/handler_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/Prescott-Data/nexus-framework/nexus-gateway/internal/broker"
+	"github.com/Prescott-Data/nexus-framework/nexus-gateway/pkg/broker"
 )
 
 // generateState creates a valid signed state string for testing
