+++ nexus-gateway/pkg/usecase/io.go
