+++ nexus-gateway/pkg/usecase/state.go
